% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheat.R
\name{cheat}
\alias{cheat}
\title{Cheat}
\usage{
cheat(exercise)
}
\arguments{
\item{exercise}{enter exercise to recall}
}
\description{
With this function you can cheat and recall your solutions to Q3.1.1, Q3.1.3, Q3.1.6, and Q3.1.16
}
\details{
Here I could even add details!
}
